chore: simplify and flatten crate::shared
