+++ src/server.rs
@@ -21,12 +21,11 @@ use lspower::lsp;
 use lspower::LanguageServer;
 
 use crate::convert;
-use crate::shared::ast::is_in_node;
-use crate::shared::get_package_name;
+use crate::shared::{get_package_name, is_in_node};
 use crate::shared::signatures::{
     get_argument_names, FunctionSignature,
 };
-use crate::shared::structs::Function;
+use crate::shared::Function;
 use crate::stdlib::{
     create_function_signature, get_builtin_functions,
     get_package_functions, get_package_infos, get_stdlib_functions,
