--- src/shared/structs.rs
@@ -1,5 +1,0 @@
-#[derive(Clone)]
-pub struct Function {
-    pub name: String,
-    pub params: Vec<String>,
-}
